+++ ext_emconf.php
@@ -27,10 +27,10 @@ $EM_CONF[$_EXTKEY] = [
     'category' => 'Base',
     'author' => 'Emily Karisch',
     'state' => 'stable',
-    'version' => '1.0.0',
+    'version' => '2.0.0',
     'constraints' => [
         'depends' => [
-            'typo3' => '8.7.0-9.5.99',
+            'typo3' => '9.5.0-10.4.99',
         ],
     ],
 ];